# backend/bait-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API in Laravel",
        "description": "API documentation with Swagger in Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/follows": {
            "post": {
                "tags": [
                    "Follows"
                ],
                "summary": "Follow another user",
                "description": "Creates a new follow relationship from the authenticated user to another user.",
                "operationId": "followUser",
                "requestBody": {
                    "description": "Pass the ID of the user to follow.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "following_id"
                                ],
                                "properties": {
                                    "following_id": {
                                        "description": "The ID of the user to be followed.",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Follow relationship created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FollowSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "409": {
                        "description": "Conflict - Already following this user"
                    },
                    "422": {
                        "description": "Unprocessable Entity (e.g., trying to follow oneself, validation error)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Follows"
                ],
                "summary": "Unfollow a user",
                "description": "Deletes an existing follow relationship.",
                "operationId": "unfollowUser",
                "requestBody": {
                    "description": "Pass the ID of the user to unfollow.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "following_id"
                                ],
                                "properties": {
                                    "following_id": {
                                        "description": "The ID of the user to be unfollowed.",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation. No content."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found - The follow relationship does not exist"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{chat}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List messages in a specific chat",
                "description": "Retrieves a paginated list of messages for a given chat. The authenticated user must be a participant of the chat.",
                "operationId": "getChatMessages",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "The ID of the chat.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessageSchema"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - User is not a participant of this chat"
                    },
                    "404": {
                        "description": "Chat not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a new message to a chat",
                "description": "Creates and broadcasts a new message in a chat. The authenticated user must be a participant of the chat.",
                "operationId": "sendMessageToChat",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "description": "The ID of the chat where the message will be sent.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The content of the message.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content_messages"
                                ],
                                "properties": {
                                    "content_messages": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "This is a test message."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - User is not a participant of this chat"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvatarSchema": {
                "title": "Avatar Schema",
                "required": [
                    "id",
                    "url_avatars"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_avatars": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/avatars/avatar1.jpg"
                    }
                },
                "type": "object"
            },
            "BannerSchema": {
                "title": "Banner Schema",
                "required": [
                    "id",
                    "url_banners"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_banners": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/banners/banner1.jpg"
                    }
                },
                "type": "object"
            },
...
            "FollowSchema": {
                "title": "Follow Schema",
                "description": "Represents a follow relationship between two users.",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the follow relationship.",
                        "type": "integer",
                        "example": 1
                    },
                    "follower": {
                        "$ref": "#/components/schemas/UserSchema"
                    },
                    "following": {
                        "$ref": "#/components/schemas/UserSchema"
                    },
                    "created_at": {
                        "description": "Timestamp when the follow action occurred.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-04T16:28:00.000000Z"
                    }
                },
                "type": "object"
            },
            "MessageSchema": {
                "title": "Message Schema",
                "description": "Represents a message within a chat.",
                "properties": {
                    "id": {
                        "description": "The message's unique identifier.",
                        "type": "integer",
                        "example": 101
                    },
                    "content_messages": {
                        "description": "The text content of the message.",
                        "type": "string",
                        "example": "Hello, how are you?"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSchema"
                    },
                    "created_at": {
                        "description": "Timestamp when the message was sent.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T12:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "UserSchema": {
                "title": "User Schema",
                "required": [
                    "id",
                    "username",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "description": "User's role name",
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "state": {
                        "description": "User's state name",
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "avatar": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AvatarSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "banner": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/BannerSchema"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-05T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a JWT token to authenticate",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}